Protect prepare-test-environment.sh against race conditions

+++ scripts/prepare-test-environment.sh
@@ -3,7 +3,7 @@
 # Shell script to initialize a pip-accel test environment.
 #
 # Author: Peter Odding <peter.odding@paylogic.com>
-# Last Change: January 17, 2016
+# Last Change: March 14, 2016
 # URL: https://github.com/paylogic/pip-accel
 #
 # This shell script is used in tox.ini and .travis.yml to prepare
@@ -48,18 +48,20 @@ install_working_copies () {
     for PROJECT in coloredlogs executor humanfriendly; do
       DIRECTORY="$HOME/projects/python/$PROJECT"
       if [ -e "$DIRECTORY" ]; then
-        msg "Installing working copy of $PROJECT .."
-        # The following code to install a Python package from a git checkout is
-        # a bit convoluted because I want to bypass pip's frustrating "let's
-        # copy the whole project tree including a 100 MB `.tox' directory
-        # before installing 10 KB of source code" approach. The use of a
-        # source distribution works fine :-).
-        cd "$DIRECTORY"
-        rm -Rf dist
-        python setup.py sdist &>/dev/null
-        # Side step caching of binary wheels because we'll be building a new
-        # one on each run anyway.
-        pip install --no-binary=:all: --quiet dist/*
+        if flock -x "$DIRECTORY"; then
+          msg "Installing working copy of $PROJECT .."
+          # The following code to install a Python package from a git checkout is
+          # a bit convoluted because I want to bypass pip's frustrating "let's
+          # copy the whole project tree including a 100 MB `.tox' directory
+          # before installing 10 KB of source code" approach. The use of a
+          # source distribution works fine :-).
+          cd "$DIRECTORY"
+          rm -Rf dist
+          python setup.py sdist &>/dev/null
+          # Side step caching of binary wheels because we'll be building a new
+          # one on each run anyway.
+          pip install --no-binary=:all: --quiet dist/*
+        fi
       fi
     done
   fi
